requirements.txt
absl-py==1.0.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
contextlib2==21.6.0
cycler==0.11.0
Cython==0.29.28
dm-env==1.5
dm-tree==0.1.6
einops==0.4.1
fasteners==0.17.3
fonttools==4.31.2
future==0.18.2
glfw==2.5.1
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gym==0.23.1
gym-notices==0.0.6
h5py==3.6.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
joblib==1.2.0
kiwisolver==1.4.2
labmaze==1.0.5
lxml==4.8.0
Markdown==3.3.6
matplotlib==3.5.1
ml-collections==0.1.1
mujoco-py==2.1.2.14
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.2
pycparser==2.21
PyOpenGL==3.1.6
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
tqdm==4.63.1
typing-extensions==4.1.1
urllib3==1.26.9
vit-pytorch==0.33.2
Werkzeug==2.1.0
zipp==3.7.0
